Fixing Hydration error
Got an error hydrartion error class:dark traced it to ThemeSwitcher
This is due to theme switching between dark and light mode.
Some java script is only applied on client side and so, it complains that the page is not static enough.
The suggestion was to add suppressHydrationWarning to the html tag